fix: #208 포크 요청 책임을 에디터 컨테이너에서 프로젝트 페이지로 변경
다른 사람의 프로젝트를 ctrl+s를 누를 때 서버에 fork 하도록 요청하는 기능은
에디터 컨테이너보다 프로젝트의 초기 데이터를 가져오고, 초기 프로젝트를 포크하는 프로젝트 페이지가
더 적합하다고 생각해 변경했습니다.
